+++ core/src/main/scala/org/apache/spark/status/api/v1/OneApplicationResource.scala
@@ -16,16 +16,150 @@
  */
 package org.apache.spark.status.api.v1
 
-import javax.ws.rs.{GET, PathParam, Produces}
-import javax.ws.rs.core.MediaType
+import java.io.OutputStream
+import java.util.{List => JList}
+import java.util.zip.ZipOutputStream
+import javax.ws.rs.{GET, Path, PathParam, Produces, QueryParam}
+import javax.ws.rs.core.{MediaType, Response, StreamingOutput}
+
+import scala.util.control.NonFatal
+
+import org.apache.spark.JobExecutionStatus
+import org.apache.spark.ui.SparkUI
 
 @Produces(Array(MediaType.APPLICATION_JSON))
-private[v1] class OneApplicationResource(uiRoot: UIRoot) {
+private[v1] class AbstractApplicationResource extends BaseAppResource {
 
   @GET
-  def getApp(@PathParam("appId") appId: String): ApplicationInfo = {
-    val apps = uiRoot.getApplicationInfo(appId)
-    apps.getOrElse(throw new NotFoundException("unknown app: " + appId))
+  @Path("jobs")
+  def jobsList(@QueryParam("status") statuses: JList[JobExecutionStatus]): Seq[JobData] = {
+    withUI(_.store.jobsList(statuses))
+  }
+
+  @GET
+  @Path("jobs/{jobId: \\d+}")
+  def oneJob(@PathParam("jobId") jobId: Int): JobData = withUI { ui =>
+    try {
+      ui.store.job(jobId)
+    } catch {
+      case _: NoSuchElementException =>
+        throw new NotFoundException("unknown job: " + jobId)
+    }
+  }
+
+  @GET
+  @Path("executors")
+  def executorList(): Seq[ExecutorSummary] = withUI(_.store.executorList(true))
+
+  @GET
+  @Path("allexecutors")
+  def allExecutorList(): Seq[ExecutorSummary] = withUI(_.store.executorList(false))
+
+  @Path("stages")
+  def stages(): Class[StagesResource] = classOf[StagesResource]
+
+  @GET
+  @Path("storage/rdd")
+  def rddList(): Seq[RDDStorageInfo] = withUI(_.store.rddList())
+
+  @GET
+  @Path("storage/rdd/{rddId: \\d+}")
+  def rddData(@PathParam("rddId") rddId: Int): RDDStorageInfo = withUI { ui =>
+    try {
+      ui.store.rdd(rddId)
+    } catch {
+      case _: NoSuchElementException =>
+        throw new NotFoundException(s"no rdd found w/ id $rddId")
+    }
+  }
+
+  @GET
+  @Path("environment")
+  def environmentInfo(): ApplicationEnvironmentInfo = withUI(_.store.environmentInfo())
+
+  @GET
+  @Path("logs")
+  @Produces(Array(MediaType.APPLICATION_OCTET_STREAM))
+  def getEventLogs(): Response = {
+    // Retrieve the UI for the application just to do access permission checks. For backwards
+    // compatibility, this code also tries with attemptId "1" if the UI without an attempt ID does
+    // not exist.
+    try {
+      withUI { _ => }
+    } catch {
+      case _: NotFoundException if attemptId == null =>
+        attemptId = "1"
+        withUI { _ => }
+        attemptId = null
+    }
+
+    try {
+      val fileName = if (attemptId != null) {
+        s"eventLogs-$appId-$attemptId.zip"
+      } else {
+        s"eventLogs-$appId.zip"
+      }
+
+      val stream = new StreamingOutput {
+        override def write(output: OutputStream): Unit = {
+          val zipStream = new ZipOutputStream(output)
+          try {
+            uiRoot.writeEventLogs(appId, Option(attemptId), zipStream)
+          } finally {
+            zipStream.close()
+          }
+
+        }
+      }
+
+      Response.ok(stream)
+        .header("Content-Disposition", s"attachment; filename=$fileName")
+        .header("Content-Type", MediaType.APPLICATION_OCTET_STREAM)
+        .build()
+    } catch {
+      case NonFatal(e) =>
+        Response.serverError()
+          .entity(s"Event logs are not available for app: $appId.")
+          .status(Response.Status.SERVICE_UNAVAILABLE)
+          .build()
+    }
+  }
+
+  /**
+   * This method needs to be last, otherwise it clashes with the paths for the above methods
+   * and causes JAX-RS to not find things.
+   */
+  @Path("{attemptId}")
+  def applicationAttempt(): Class[OneApplicationAttemptResource] = {
+    if (attemptId != null) {
+      throw new NotFoundException(httpRequest.getRequestURI())
+    }
+    classOf[OneApplicationAttemptResource]
+  }
+
+}
+
+private[v1] class OneApplicationResource extends AbstractApplicationResource {
+
+  @GET
+  def getApp(): ApplicationInfo = {
+    val app = uiRoot.getApplicationInfo(appId)
+    app.getOrElse(throw new NotFoundException("unknown app: " + appId))
+  }
+
+}
+
+private[v1] class OneApplicationAttemptResource extends AbstractApplicationResource {
+
+  @GET
+  def getAttempt(): ApplicationAttemptInfo = {
+    uiRoot.getApplicationInfo(appId)
+      .flatMap { app =>
+        app.attempts.filter(_.attemptId == attemptId).headOption
+      }
+      .getOrElse {
+        throw new NotFoundException(s"unknown app $appId, attempt $attemptId")
+      }
   }
 
 }
